requirements.txt
selenium==4.8.2
Appium_Python_Client==2.8.1
pytest==7.2.2
allure-pytest==2.13.1
tox==4.4.7
allure-python-commons~=2.13.1
webdriver_manager==4.0.0